[NFC] Split common.h header [1/N]
This PR outlines functions like `linearize` and `unlinearize` to their
own separate header, because they are only used by a few tests.

The change is expected to have positive (albeit probably unnoticeable)
impact on overall CTS compile-times.

The main motivation/background is to refactor CTS in preparation for
using modularized `sycl.hpp` (see KhronosGroup/SYCL-Docs#780). As of
today all tests use `common.h` and therefore CTS won't benefit much from
fine grained headers because everything anyways is included through a
single entry point in form of `common.h`.
