+++ tests/extension/oneapi_auto_local_range/auto_local_range.cpp
@@ -19,6 +19,7 @@
 *******************************************************************************/
 
 #include "../../common/common.h"
+#include "../../common/index_space.h"
 #include <type_traits>
 
 namespace auto_local_range::tests {
