+++ tests/host_accessor/host_accessor_api_common.h
@@ -8,6 +8,7 @@
 #ifndef SYCL_CTS_HOST_ACCESSOR_API_COMMON_H
 #define SYCL_CTS_HOST_ACCESSOR_API_COMMON_H
 #include "../accessor_basic/accessor_common.h"
+#include "../common/index_space.h"
 #include <cmath>
 
 namespace host_accessor_api_common {
